build: fix prettier ignore
In previous commits, we forgot to include the .prettierignore file as
the --ingore-path option was already used. As prettier doesn't support
multiple ignore files for now, we therefore need to duplicate .gitignore
in .prettierignore. Here, I used ignore-sync to automate this and avoid
any future omission, but still included the generated .prettierignore to
ease contribution.