build: only build raw on watch
Rollup fails to rewrite raw build files when using watch mode, while
bundled files are overwritten correctly. This seems to be linked to
Rollup having issues when running watch mode with several build configs.
As a workaround, the npm watch script now only runs the first "raw"
build in the configuration (the one for npm). This should be enough as
demos and storybook only use this build, while the bundles
are only used in the static demo (wich require to rerun the bundle
script anyway) or when loading them from the unpkgr CDN (and well, we
can't have a watch on this).